+++ HISTORY.rst
@@ -3,6 +3,18 @@
 History
 -------
 
+3.0.0 (2015-12-01)
+++++++++++++++++++
+
+* BREAKING CHANGE: client.bundle_list_map(func) The func is now called
+with client as the first parameter.
+* Added insight access methods.
+* Support for requesting non-autorun insights (ex. transcripts and captions)
+* Support for embedding insights when requesting bundles.
+* Support for external_id when creating or updating bundles.
+* Added behave tests.
+* Added unittests.
+
 2.0.0 (2015-04-18)
 ++++++++++++++++++